readding argon + moving jskeymanager to util
