+++ package.json
@@ -54,6 +54,7 @@
   },
   "dependencies": {
     "@greymass/eosio": "^0.6.2",
+    "argon2": "^0.29.1",
     "cross-fetch": "^3.1.5",
     "randombytes": "^2.1.0",
     "scrypt-js": "^3.0.1"
