+++ src/keymanager.ts
@@ -1,5 +1,4 @@
-import { PrivateKey, PublicKey, Checksum256, Signature } from '@greymass/eosio';
-import { randomString, sha256 } from './util/crypto';
+import { PrivateKey, PublicKey, Signature } from '@greymass/eosio';
 
 enum KeyManagerLevel {
     PASSWORD = 'PASSWORD',
@@ -94,6 +93,22 @@ interface KeyManager {
      * @throws if a key does not exist for the level
      */
     getKey(options: GetKeyOptions): PublicKey
+
+    /**
+     * generates a random private key
+     *  
+     * @returns The PrivateKey
+     */
+    generateRandomPrivateKey(): PrivateKey;
+
+    /**
+     * generates a private key from a password and creates random salt
+     * @param password password to encrypt the private key with
+     * @returns encrypted private key and salt
+     * 
+     */
+    generatePrivateKeyFromPassword(password: string): Promise<{ privateKey: PrivateKey, salt: Buffer }>;
+
 }
 
 type KeyStorage = {
@@ -104,59 +119,6 @@ type KeyStorage = {
     salt?: string;
 }
 
-class JsKeyManager implements KeyManager {
-    // TODO: use localStorage or sessionStorage in browser if available instead of keyStorage
-    keyStorage: {
-        [key in KeyManagerLevel]: KeyStorage;
-    }
 
 
-
-
-    // creates a new secure key and returns the public key
-    async storeKey(options: StoreKeyOptions): Promise<PublicKey> {
-        const keyStore: KeyStorage = {
-            privateKey: options.privateKey,
-            publicKey: options.privateKey.toPublic()
-        }
-
-        if (options.level === KeyManagerLevel.PASSWORD || options.level === KeyManagerLevel.PIN) {
-            if (!options.challenge) throw new Error("Challenge missing");
-
-            keyStore.salt = randomString(32);
-            keyStore.hashedSaltedChallenge = sha256(options.challenge + keyStore.salt);
-        }
-
-        this.keyStorage[options.level] = keyStore;
-        return keyStore.publicKey;
-    }
-
-    async signData(options: SignDataOptions): Promise<string | Signature> {
-        if (options.level in this.keyStorage) throw new Error("No key for this level");
-
-        const keyStore = this.keyStorage[options.level];
-
-        if (options.level === KeyManagerLevel.PASSWORD || options.level === KeyManagerLevel.PIN) {
-            if (!options.challenge) throw new Error("Challenge missing");
-
-            const hashedSaltedChallenge = sha256(options.challenge + keyStore.salt);
-
-            if (keyStore.hashedSaltedChallenge !== hashedSaltedChallenge) throw new Error("Challenge does not match");
-        }
-
-        const privateKey = keyStore.privateKey;
-        const hash = Checksum256.hash(options.data);
-        const signature = privateKey.signDigest(hash)
-
-        return signature;
-    }
-
-    getKey(options: GetKeyOptions): PublicKey {
-        if (options.level in this.keyStorage) throw new Error("No key for this level");
-
-        const keyStore = this.keyStorage[options.level];
-        return keyStore.publicKey;
-    }
-}
-
-export { KeyManager, KeyManagerLevel, JsKeyManager }
+export { KeyManager, KeyManagerLevel, KeyStorage, StoreKeyOptions, SignDataOptions, GetKeyOptions };
